- Fix runtime library order - Add default OpenGL library on Windows from Angle - Dynamic bind GLES2 and EGL to default rendering functions - Use EGLNativeWindowType for generic managment

+++ src/videofilters/nowebcam.c
@@ -380,7 +380,7 @@ void ms_static_image_set_default_image(const char *path){
 	if (ms_nowebcam_def_image[0] != 0)
 		ms_nowebcam_def_image[0] = 0;
 	if (path) {
-		int min = MIN(DEF_IMAGE_MAX_SIZE, strlen(path));
+		int min = MIN(DEF_IMAGE_MAX_SIZE, (int)strlen(path));
 		strncpy(ms_nowebcam_def_image, path, min);
 		ms_nowebcam_def_image[min] = '\0';
 	}
